remove ints from syscalls, improve remove, add errnos

+++ headers/fcntl.h
@@ -1,6 +1,6 @@
 #ifndef __FCNTL_H
 #define __FCNTL_H
-int open(const char *, int, mode_t);
+long open(const char *, int, mode_t);
 #define __first(a, ...) a
 #define open(path, flags, ...) open(path, flags, __first(__VA_ARGS__+0))
